Improving data id types, and swagger notation

+++ RestAspNet/Business/IBookBusiness.cs
@@ -6,7 +6,7 @@ namespace RestAspNet.Business
     public interface IBookBusiness
     {
         BookVO Create(BookVO book);
-        BookVO FindByID(long id);
+        BookVO FindByID(int id);
         List<BookVO> FindAll();
         BookVO Update(BookVO book);
         void Delete(long id);
