+++ RestAspNet/Repository/Generic/IRepository.cs
@@ -5,7 +5,7 @@ namespace RestAspNet.Repository.Generic
     public interface IRepository<T> where T : BaseEntity
     {
         T Create(T item);
-        T FindByID(long id);
+        T FindByID(int id);
         List<T> FindAll();
         T Update(T item);
         void Delete(long id);
